add mock data to shared file
