+++ web/src/stories/mockData.ts
@@ -124,3 +124,129 @@ export const zoneStateMock: StateZoneData = {
   },
   e: 0,
 };
+
+export const priceData = {
+  entryCount: 24,
+  priceData: {
+    '2024-09-01 22:00:00+00:00': 25,
+    '2024-09-01 23:00:00+00:00': 15,
+    '2024-09-02 00:00:00+00:00': 12,
+    '2024-09-02 01:00:00+00:00': 8,
+    '2024-09-02 02:00:00+00:00': 21,
+    '2024-09-02 03:00:00+00:00': 16,
+    '2024-09-02 04:00:00+00:00': 19,
+    '2024-09-02 05:00:00+00:00': 24,
+    '2024-09-02 06:00:00+00:00': 27,
+    '2024-09-02 07:00:00+00:00': 22,
+    '2024-09-02 08:00:00+00:00': 13,
+    '2024-09-02 09:00:00+00:00': 29,
+    '2024-09-02 10:00:00+00:00': 18,
+    '2024-09-02 11:00:00+00:00': 20,
+    '2024-09-02 12:00:00+00:00': 26,
+    '2024-09-02 13:00:00+00:00': 14,
+    '2024-09-02 14:00:00+00:00': 23,
+    '2024-09-02 15:00:00+00:00': 30,
+    '2024-09-02 16:00:00+00:00': 17,
+    '2024-09-02 17:00:00+00:00': 11,
+    '2024-09-02 18:00:00+00:00': 22,
+    '2024-09-02 19:00:00+00:00': 13,
+    '2024-09-02 20:00:00+00:00': 28,
+    '2024-09-02 21:00:00+00:00': 19,
+    '2024-09-02 22:00:00+00:00': 29,
+    '2024-09-02 23:00:00+00:00': 16,
+    '2024-09-03 00:00:00+00:00': 24,
+    '2024-09-03 01:00:00+00:00': 15,
+  },
+  currency: 'EUR',
+  source: 'nordpool.com',
+  zoneKey: 'DE',
+};
+
+export const halfHourPriceData = {
+  entryCount: 28,
+  priceData: {
+    '2024-09-01 22:00:00+00:00': 25,
+    '2024-09-01 22:30:00+00:00': 15,
+    '2024-09-01 23:00:00+00:00': 12,
+    '2024-09-01 23:30:00+00:00': 28,
+    '2024-09-02 00:00:00+00:00': 21,
+    '2024-09-02 00:30:00+00:00': -5,
+    '2024-09-02 01:00:00+00:00': 19,
+    '2024-09-02 01:30:00+00:00': 24,
+    '2024-09-02 02:00:00+00:00': 27,
+    '2024-09-02 02:30:00+00:00': 22,
+    '2024-09-02 03:00:00+00:00': -10,
+    '2024-09-02 03:30:00+00:00': 29,
+    '2024-09-02 04:00:00+00:00': 18,
+    '2024-09-02 04:30:00+00:00': 20,
+    '2024-09-02 05:00:00+00:00': 26,
+    '2024-09-02 05:30:00+00:00': -3,
+    '2024-09-02 06:00:00+00:00': 23,
+    '2024-09-02 06:30:00+00:00': 40,
+    '2024-09-02 07:00:00+00:00': 17,
+    '2024-09-02 07:30:00+00:00': 11,
+    '2024-09-02 08:00:00+00:00': 22,
+    '2024-09-02 08:30:00+00:00': 13,
+    '2024-09-02 09:00:00+00:00': 28,
+    '2024-09-02 09:30:00+00:00': 19,
+    '2024-09-02 10:00:00+00:00': 29,
+    '2024-09-02 10:30:00+00:00': 16,
+    '2024-09-02 11:00:00+00:00': 24,
+    '2024-09-02 11:30:00+00:00': 15,
+    '2024-09-02 12:00:00+00:00': 12,
+    '2024-09-02 12:30:00+00:00': 28,
+    '2024-09-02 13:00:00+00:00': 21,
+    '2024-09-02 13:30:00+00:00': -5,
+    '2024-09-02 14:00:00+00:00': 19,
+    '2024-09-02 14:30:00+00:00': 24,
+    '2024-09-02 15:00:00+00:00': 27,
+    '2024-09-02 15:30:00+00:00': 22,
+    '2024-09-02 16:00:00+00:00': -10,
+    '2024-09-02 16:30:00+00:00': 29,
+    '2024-09-02 17:00:00+00:00': 18,
+    '2024-09-02 17:30:00+00:00': 20,
+    '2024-09-02 18:00:00+00:00': 26,
+    '2024-09-02 18:30:00+00:00': -3,
+    '2024-09-02 19:00:00+00:00': 23,
+    '2024-09-02 19:30:00+00:00': 30,
+    '2024-09-02 20:00:00+00:00': 17,
+    '2024-09-02 20:30:00+00:00': 11,
+    '2024-09-02 21:00:00+00:00': 22,
+    '2024-09-02 21:30:00+00:00': 13,
+    '2024-09-02 22:00:00+00:00': 28,
+    '2024-09-02 22:30:00+00:00': 19,
+    '2024-09-02 23:00:00+00:00': 29,
+    '2024-09-02 23:30:00+00:00': 16,
+    '2024-09-03 00:00:00+00:00': 24,
+    '2024-09-03 00:30:00+00:00': 15,
+    '2024-09-03 01:00:00+00:00': 12,
+    '2024-09-03 01:30:00+00:00': 28,
+    '2024-09-03 02:00:00+00:00': 21,
+    '2024-09-03 02:30:00+00:00': -5,
+    '2024-09-03 03:00:00+00:00': 19,
+    '2024-09-03 03:30:00+00:00': 24,
+    '2024-09-03 04:00:00+00:00': 27,
+    '2024-09-03 04:30:00+00:00': 22,
+    '2024-09-03 05:00:00+00:00': -10,
+    '2024-09-03 05:30:00+00:00': 29,
+    '2024-09-03 06:00:00+00:00': 18,
+    '2024-09-03 06:30:00+00:00': 20,
+    '2024-09-03 07:00:00+00:00': 26,
+    '2024-09-03 07:30:00+00:00': -3,
+    '2024-09-03 08:00:00+00:00': 23,
+    '2024-09-03 08:30:00+00:00': 30,
+    '2024-09-03 09:00:00+00:00': 17,
+    '2024-09-03 09:30:00+00:00': 11,
+    '2024-09-03 10:00:00+00:00': 22,
+    '2024-09-03 10:30:00+00:00': 13,
+    '2024-09-03 11:00:00+00:00': 28,
+    '2024-09-03 11:30:00+00:00': 19,
+    '2024-09-03 12:00:00+00:00': 29,
+    '2024-09-03 12:30:00+00:00': 16,
+    '2024-09-03 13:00:00+00:00': 24,
+    '2024-09-03 13:30:00+00:00': 15,
+  },
+  currency: 'GBP',
+  source: 'nordpool.com',
+  zoneKey: 'GB',
+};
